refactor(MEP): move MEP headers into src/include/mep
